{{ cookiecutter.directory_name }}/requirements.txt
Ipython
dj-database-url==0.3.0
django-bootstrap3==7.0.0
django-debug-toolbar==1.4
django-extensions==1.7.4
django-hamlpy==0.86
django-nose==1.4.4
django-pipeline==1.6.8
django==1.10.3
factory-boy==2.7.0
flake8
gunicorn==19.4.5
isort
newrelic==2.60.0.46
pep8
psycopg2==2.6.1
pydocstyle
python-dotenv==0.6.1
tqdm==4.7.1
whitenoise==2.0.6
